+++ src/hooks/transactions/helpers/timestampIsOlderThan.ts
@@ -1,0 +1,7 @@
+export const timestampIsOlderThan = (
+  timestamp: number,
+  olderThanMs: number
+) => {
+  const diff = new Date().getTime() - timestamp;
+  return diff > olderThanMs;
+};